Split test_api.py into a new folder with files for each endpoint to test
