+++ tests/test_api/test_items.py
@@ -3,12 +3,50 @@ from datetime import datetime
 import dateutil.parser
 
 class TestItems:
-    def test_get_item(self, app):
+    def test_get_empty_items(self, app):
         client = app.test_client()
 
         response = client.get('/items/')
         assert response.status_code == 200
         assert response.json == []
+
+    def test_get_nonempty_items(self, app):
+        client = app.test_client()
+
+        new_item = {
+            'name':'Pineapple',
+            'amount':2,
+        }
+
+        response = client.post('/items/', 
+            headers = {"Content-Type": "application/json"},
+            data = json.dumps(new_item),
+        )
+
+        response = client.get('/items/')
+        assert response.status_code == 200
+        assert len(response.json) == 1
+
+    def test_get_invalid_item(self, app):
+        client = app.test_client()
+
+        new_item = {
+            'name':'Pineapple',
+            'amount':2,
+        }
+
+        response = client.post('/items/', 
+            headers = {"Content-Type": "application/json"},
+            data = json.dumps(new_item),
+        )
+
+        assert response.status_code == 201
+
+        id = response.json['id']
+
+        response = client.get(f'/items/{id+1}')
+
+        assert response.status_code == 404
 
     def test_post_item(self, app):
         client = app.test_client()
@@ -33,6 +71,29 @@ class TestItems:
         assert response.status_code == 200
         for k,v in new_item.items():
             assert response.json[k] == v
+
+    def test_post_existing_item(self, app):
+        client = app.test_client()
+
+        new_item = {
+            'name':'Kroger Eggs',
+            'amount':12,
+            'product_id':123456,
+        }
+
+        response = client.post('/items/', 
+            headers = {"Content-Type": "application/json"},
+            data = json.dumps(new_item),
+        )
+
+        assert response.status_code == 201
+
+        response = client.post('/items/', 
+            headers = {"Content-Type": "application/json"},
+            data = json.dumps(new_item),
+        )
+
+        assert response.status_code == 422
 
     def test_put_item(self, app):
         client = app.test_client()
@@ -229,7 +290,6 @@ class TestItems:
         assert response.status_code == 200
         assert len(response.json) == 2
 
-
     def test_post_invalid_item(self, app):
         client = app.test_client()
 
@@ -319,130 +379,3 @@ class TestItems:
         )
 
         assert response.status_code == 404
-
-class TestIngredients:
-    def test_get_empty_ingredients(self, app):
-        client = app.test_client()
-
-        response = client.get('/ingredients/')
-        assert response.status_code == 200
-        assert response.json == []
-
-    def test_post_get_ingredient(self, app):
-        client = app.test_client()
-
-        new_ingredient = {
-            'name':'Eggs',
-        }
-
-        response = client.post('/ingredients/', 
-            headers = {"Content-Type": "application/json"},
-            data = json.dumps(new_ingredient),
-        )
-
-        assert response.status_code == 201
-
-        id = response.json['id']
-
-        response = client.get(f'/ingredients/{id}')
-
-        assert response.status_code == 200
-        for k,v in new_ingredient.items():
-            assert response.json[k] == v
-
-    def test_post_invalid_ingredient(self, app):
-        client = app.test_client()
-
-        # missing name
-        new_ingredient = {
-            
-        }
-
-        response = client.post('/ingredients/', 
-            headers = {"Content-Type": "application/json"},
-            data = json.dumps(new_ingredient),
-        )
-
-        assert response.status_code == 422
-
-    def test_delete_ingredient(self, app):
-        client = app.test_client()
-
-        new_ingredient = {
-            'name':'Spinach',
-        }
-
-        response = client.post('/ingredients/', 
-            headers = {"Content-Type": "application/json"},
-            data = json.dumps(new_ingredient),
-        )
-
-        assert response.status_code == 201
-
-        id = response.json['id']
-        etag = response.headers['ETag']
-
-        response = client.delete(f'/ingredients/{id}',
-            headers={'If-Match': etag}
-        )
-
-        assert response.status_code == 204
-
-class TestItemIngredients:
-    def test_link_item_ingredient(self, app):
-        client = app.test_client()
-
-        # create an item with no ingredient id set
-        new_item1 = {
-            'name':'Kroger Chicken Thighs',
-            'amount':6,
-            'product_id':1111111111111,
-        }
-
-        response = client.post('/items/', 
-            headers = {"Content-Type": "application/json"},
-            json = new_item1,
-        )
-
-        assert response.status_code == 201
-
-        item1_id = response.json['id']
-        item1_etag = response.headers['ETag']
-
-        # create an ingredient
-        new_ingredient = {
-            'name':'chicken thighs',
-        }
-
-        response = client.post('/ingredients/', 
-            headers = {"Content-Type": "application/json"},
-            json = new_ingredient,
-        )
-
-        assert response.status_code == 201
-
-        ingredient_id = response.json['id']
-
-        # create another new item, this time with the newly created ingredient id
-        new_item2 = {
-            'name':'Costco Chicken Thighs',
-            'amount':2,
-            'product_id':2222222222222,
-        }
-
-        response = client.post(f'/items/', 
-            headers = {"Content-Type": "application/json"},
-            json = new_item2,
-        )
-
-        assert response.status_code == 201
-
-        # update the first item to use the ingredient
-        new_item1['ingredient_id'] = ingredient_id
-        response = client.put(f'/items/{item1_id}', 
-            headers = {"If-Match": item1_etag},
-            json = new_item1,
-        )
-
-        assert response.status_code == 200
-        assert response.json['ingredient_id'] == ingredient_id
